Update painel.sh
Agora é checado se os arquivos criado pelo script já existem, e dá a opção de fazer uma nova configuração;
Também corrigido um problema aonde não era impedido de digitar entradas erradas durante o script.

+++ painel.sh
@@ -1,15 +1,36 @@
 #!/usr/bin/env bash
 
+# Verificar se existe o diretório chinchila para seguir ou não com o script
+if test -e "/home/alpha7/chinchila-online-arquivos/"; then
+	while [[ ! "$refaz1" =~ [snSN] ]]; do
+	echo -n "O arquivo ou diretório já existe, deseja proseguir com o script? [S/N] "
+	read refaz1
+	done
+
+	if [[ "$refaz1" =~ [sS] ]]; then
+		rm -r /home/alpha7/chinchila-online-arquivos/ && rm -r /home/alpha7/imprimeEtiquetaAtendimento
+	fi
+
+	if [[ "$refaz1" =~ [nN] ]]; then
+		exit 0
+	fi
+fi
+# Caso tenha selecionado S, será dado sequência com o script do painel
 cd /home/alpha7
 mkdir chinchila-online-arquivos && cd chinchila-online-arquivos
 wget -q https://cdn.discordapp.com/attachments/981227657595330623/1029111840929746974/logo_a7.png
-echo -n "=================================================================
+# Selecionar o tema do painel de senhas, sempre é bom verificar com o cliente qual a rede do mesmo
+while [[ ! "$css" =~ [1234] ]]; do
+	echo -n "=================================================================
 (1)Padrão
 (2)Maxi Popular
 (3)Ultra Popular
 (4)Escuro
 Baixar Painel: "
 read css
+done
+# Como o arquivo CSS não é de minha autoria não deixei o mesmo salvo no repositório
+# De alguma forma os links do discord duram por muito tempo, optei por então usar eles para baixar o arquivo css
 case $css in
 	1)
 		echo "Baixado Painel Padrão"
@@ -31,13 +52,17 @@ esac
 cd -
 touch imprimeEtiquetaAtendimento
 chmod u+x imprimeEtiquetaAtendimento
+# Configurar em como será feito a impressão do cupom de atendimento, este pode ser feito por tablet e computador
+# Caso seja um tablet windows pode ser usado a opção de computador
 echo "=================================================================
 Dado Permissão de Execução ao Arquivo: imprimeEtiquetaAtendimento"
+while [[ ! "$var" =~ [12] ]]; do
 echo -n "=================================================================
 (1)Tablet
 (2)Computador
 Será configurado: "
 read var
+done
 case $var in
 	1)
 		echo "================================================================="
@@ -61,16 +86,15 @@ case $var in
 		echo "smbclient //$IP_COMPUTADOR_INSTALACAO/$NOME_COMPARTILHAMENTO_IMPRESSORA -U "\"$USUARIO%$SENHA\"" -c "\"print "\${1}\"" >> imprimeEtiquetaAtendimento
 		;;
 esac
+# Finalização da configuração do painel de senhas, caso preenchido alguma informação incorreta pode ser cancelado o mesmo
+while [[ ! "$refaz" =~ [snSN] ]]; do
 echo -n "=================================================================
-(1) Sim
-(2) Não
-Desfazer configuração do Painel de Senhas? "
+Desfazer configuração do Painel de Senhas? [S/N] "
 read refaz
-case $refaz in
-	1)
-		rm -r chinchila-online-arquivos && rm imprimeEtiquetaAtendimento
-		;;
-	2)
-		echo "Configuração Completa"
-	 	;;
-esac
+done
+if [[ "$refaz" =~ [sS] ]]; then
+	rm -r chinchila-online-arquivos && rm imprimeEtiquetaAtendimento
+fi
+if [[ "$refaz" =~ [nN] ]]; then
+	echo "Configuração Completa"
+fi
